Removing dependencies pertaining to prophet
