+++ Dockerfile
@@ -1,5 +1,6 @@
 FROM registry.access.redhat.com/ubi8/python-38:latest
 
+
 # Add application sources to a directory that the assemble script expects them
 # and set permissions so that the container runs without root access
 USER 0
