+++ Pipfile
@@ -10,7 +10,6 @@ prometheus_client = "*"
 prometheus_api_client = "*"
 tornado = "*"
 pystan = "==2.19.1.1"
-prophet = "*"
 dateparser = "*"
 mlflow = "*"
 numpy = "*"
