+++ environment.yml
@@ -5,7 +5,6 @@ channels:
 dependencies:
   - python=3.8
   - pip
-  - prophet
   - pip:
       - pandas
       - schedule
