+++ requirements.txt
@@ -4,7 +4,6 @@ prometheus_client
 prometheus_api_client
 tornado
 pystan==2.19.1.1
-prophet
 mlflow
 numpy
 scikit-learn
